% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ni_clinical_loneliness_sdz}
\alias{ni_clinical_loneliness_sdz}
\title{Northern Ireland Loneliness Scores from GP prescription data, 2022}
\format{
A data frame with 850 rows and 4 variables:
\describe{
\item{sdz21_code}{Super Data Zone Code}
\item{loneliness_zscore}{Loneliness score}
\item{rank}{Rank of loneliness score}
\item{deciles}{Loneliness score deciled}
...
}
}
\source{
\url{https://www.data.gov.uk}
}
\usage{
ni_clinical_loneliness_sdz
}
\description{
A dataset with scores for loneliness at SDZ level.
The approach and code to develop this index is based on the
Office for National Statistics' Data Science Campus Loneliness Prescription Index.
}
\keyword{datasets}
