% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotland_clinical_loneliness_dz}
\alias{scotland_clinical_loneliness_dz}
\title{Scotland Loneliness Scores from GP prescription data, 2022}
\format{
A data frame with 6,976 rows and 4 variables:
\describe{
\item{dz11_code}{Intermediate Zone Code}
\item{loneliness_zscore}{Loneliness score}
\item{rank}{Rank of loneliness score}
\item{deciles}{Loneliness score deciled - 1 is least lonely, 10 is most lonely}
...
}
}
\source{
\url{https://www.opendata.nhs.scot}
}
\usage{
scotland_clinical_loneliness_dz
}
\description{
A dataset with scores for loneliness at DZ level based on GP prescription data.
The approach and code to develop this index is based on the
Office for National Statistics' Data Science Campus Loneliness Prescription Index.
Data is not available for 172 DataZones (2\% of total DataZones) due to lack of
proximity to GPs. Their scores are marked as NA.
}
\keyword{datasets}
