% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wales_clinical_loneliness_lsoa}
\alias{wales_clinical_loneliness_lsoa}
\title{Wales Loneliness Scores from GP prescription data, 2022}
\format{
A data frame with 1,917 rows and 4 variables:
\describe{
\item{lsoa21_code}{Lower Super Output Area Code}
\item{loneliness_zscore}{Loneliness score}
\item{rank}{Rank of loneliness score}
\item{deciles}{Loneliness score deciled}
...
}
}
\source{
\url{https://nwssp.nhs.wales/}
}
\usage{
wales_clinical_loneliness_lsoa
}
\description{
A dataset with scores for loneliness at LSOA level based on GP prescription data.
The approach and code to develop this index is based on the
Office for National Statistics' Data Science Campus Loneliness Prescription Index.
}
\keyword{datasets}
