% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_cls_loneliness_2020}
\alias{england_cls_loneliness_2020}
\title{England Loneliness Scores from Community Life Survey 2020-21}
\format{
A data frame with 33,755 rows and 2 variables:
\describe{
\item{lsoa21_code}{Lower Super Output Area 2021 Code}
\item{perc}{Perc}
...
}
}
\source{
\url{...}
}
\usage{
england_cls_loneliness_2020
}
\description{
A dataset...
...
}
\keyword{datasets}
